Fix modifier keys stuck down after unfocusing the window
