+++ embedder/Cargo.toml
@@ -9,7 +9,7 @@ members = [
 ]
 
 [dependencies]
-smithay = { git = "https://github.com/Smithay/smithay", rev = "123f790", features = ["default", "wayland_frontend", "backend_egl", "use_system_lib"] }
+smithay = { git = "https://github.com/roscale/smithay", rev = "92acbe5", features = ["default", "wayland_frontend", "backend_egl", "use_system_lib"] }
 smithay-drm-extras = {path = "third_party/smithay-drm-extras" }
 rustix = "0.38.30"
 input-linux = "0.6.0"
